+++ mle_toolbox/remote/gcloud_transfer.py
@@ -8,7 +8,7 @@ import shutil
 from datetime import datetime
 from typing import Union
 
-from ..utils.general import load_mle_toolbox_config
+from ..utils import load_mle_toolbox_config
 from ..protocol import load_local_protocol_db
 from .ssh_transfer import setup_proxy_server
 
@@ -32,18 +32,19 @@ def get_gcloud_db(number_of_connect_tries: int=5):
             blob = bucket.blob(cc.gcp.protocol_fname)
             with open(expanduser(cc.general.local_protocol_fname), 'wb') as file_obj:
                 blob.download_to_file(file_obj)
-            logger.info("Pulled from GCloud Storage - {}".format(cc.gcp.protocol_fname))
+            logger.info(f"Pulled from GCloud Storage - {cc.gcp.protocol_fname}")
             return 1
         except Exception as ex:
             if type(ex).__name__ == "NotFound":
-                logger.info("No DB found in GCloud Storage - {}".format(cc.gcp.protocol_fname))
-                logger.info("New DB will be created - {}/{}".format(cc.gcp.project_name,
-                                                                    cc.gcp.bucket_name))
+                logger.info(f"No DB found in GCloud Storage"
+                            + f" - {cc.gcp.protocol_fname}")
+                logger.info("New DB will be created - "
+                            + f"{cc.gcp.project_name}/{cc.gcp.bucket_name}")
                 return 1
             else:
-                logger.info("Attempt {}/{} - Failed pulling from GCloud Storage - {}".format(i+1,
-                                                                                             number_of_connect_tries,
-                                                                                             type(ex).__name__))
+                logger.info(f"Attempt {i+1}/{number_of_connect_tries}"
+                            + f" - Failed pulling from GCloud Storage"
+                            + f" - {type(ex).__name__}")
     # If after 5 pulls no successful connection established - return failure
     return 0
 
@@ -58,12 +59,13 @@ def send_gcloud_db(number_of_connect_tries: int=5):
             client = storage.Client(cc.gcp.project_name)
             bucket = client.get_bucket(cc.gcp.bucket_name, timeout=20)
             blob = bucket.blob(cc.gcp.protocol_fname)
-            blob.upload_from_filename(filename=expanduser(cc.general.local_protocol_fname))
-            logger.info("Send to GCloud Storage - {}".format(cc.gcp.protocol_fname))
+            blob.upload_from_filename(
+                filename=expanduser(cc.general.local_protocol_fname))
+            logger.info(f"Send to GCloud Storage - {cc.gcp.protocol_fname}")
             return 1
         except:
-            logger.info("Attempt {}/{} - Failed sending to GCloud Storage".format(i+1,
-                                                                                  number_of_connect_tries))
+            logger.info(f"Attempt {i+1}/{number_of_connect_tries}"
+                        + f" - Failed sending to GCloud Storage")
     # If after 5 pulls no successful connection established - return failure
     return 0
 
@@ -78,7 +80,8 @@ def delete_gcs_directory(gcs_path: str,
             client = storage.Client(cc.gcp.project_name)
             bucket = client.get_bucket(cc.gcp.bucket_name, timeout=20)
         except:
-            logger.info(f"Attempt {i+1}/{number_of_connect_tries} - Failed sending to GCloud Storage")
+            logger.info(f"Attempt {i+1}/{number_of_connect_tries}"
+                        + f" - Failed sending to GCloud Storage")
 
     # Delete all files in directory
     blobs = bucket.list_blobs(prefix=gcs_path)
@@ -96,16 +99,22 @@ def upload_local_directory_to_gcs(local_path: str, gcs_path: str,
             client = storage.Client(cc.gcp.project_name)
             bucket = client.get_bucket(cc.gcp.bucket_name, timeout=20)
         except:
-            logger.info(f"Attempt {i+1}/{number_of_connect_tries} - Failed sending to GCloud Storage")
+            logger.info(f"Attempt {i+1}/{number_of_connect_tries}"
+                        + f" - Failed sending to GCloud Storage")
 
     def upload_single_file(local_path, gcs_path, bucket):
         # Recursively upload the folder structure
         if os.path.isdir(local_path):
             for local_file in glob.glob(os.path.join(local_path, '**')):
                 if not os.path.isfile(local_file):
-                   upload_single_file(local_file, os.path.join(gcs_path, os.path.basename(local_file)), bucket)
+                   upload_single_file(local_file,
+                                      os.path.join(
+                                        gcs_path,
+                                        os.path.basename(local_file)),
+                                      bucket)
                 else:
-                   remote_path = os.path.join(gcs_path, local_file[1 + len(local_path):])
+                   remote_path = os.path.join(gcs_path,
+                                              local_file[1 + len(local_path):])
                    blob = bucket.blob(remote_path)
                    blob.upload_from_filename(local_file)
         # Only upload single file - e.g. zip compressed experiment
@@ -127,7 +136,8 @@ def download_gcs_directory_to_local(gcs_path: str, local_path: str="",
             client = storage.Client(cc.gcp.project_name)
             bucket = client.get_bucket(cc.gcp.bucket_name, timeout=20)
         except:
-            logger.info(f"Attempt {i+1}/{number_of_connect_tries} - Failed sending to GCloud Storage")
+            logger.info(f"Attempt {i+1}/{number_of_connect_tries}"
+                        + f" - Failed sending to GCloud Storage")
 
     blobs = bucket.list_blobs(prefix=gcs_path)  # Get list of files
     blobs = list(blobs)
@@ -208,7 +218,8 @@ def get_gcloud_zip_experiment(db, experiment_id: str,
 
         if experiment_id not in all_experiment_ids:
             print(time_t, "The experiment you try to retrieve does not exist")
-            experiment_id = input(time_t + " Which experiment do you want to retrieve?")
+            experiment_id = input(time_t +
+                                  " Which experiment do you want to retrieve?")
         else:
             break
 
@@ -231,7 +242,8 @@ def get_gcloud_zip_experiment(db, experiment_id: str,
 
     time_t = datetime.now().strftime("%m/%d/%Y %I:%M:%S %p")
     # Goodbye message if successful
-    print(time_t, f"Successfully retrieved {experiment_id} - from GCS {gcloud_hash_fname}")
+    print(time_t, f"Successfully retrieved {experiment_id}"
+                  + f" - from GCS {gcloud_hash_fname}")
 
     # Update protocol retrieval status of the experiment
     db.dadd(experiment_id, ("retrieved_results", True))