+++ mle_toolbox/remote/ssh_transfer.py
@@ -3,7 +3,7 @@ import paramiko
 from scp import SCPClient
 from sshtunnel import SSHTunnelForwarder
 from os.path import expanduser
-from ..utils.general import determine_resource, load_mle_toolbox_config
+from ..utils import determine_resource, load_mle_toolbox_config
 
 
 def setup_proxy_server():