+++ mle_toolbox/utils/__init__.py
@@ -1,16 +1,19 @@
-from .general import (load_mle_toolbox_config,
-                      load_yaml_config,
-                      load_json_config,
-                      load_pkl_object,
-                      determine_resource,
-                      print_framed,
-                      get_jax_os_ready)
+from .core_experiment import (get_configs_ready,
+                              set_random_seeds,
+                              determine_resource)
+from .core_files_load import (load_mle_toolbox_config,
+                              load_yaml_config,
+                              load_json_config,
+                              load_pkl_object,
+                              load_result_logs)
+from .core_files_merge import merge_hdf5_files
 from .mle_logger import MLE_Logger
-from .core import get_configs_ready, set_random_seeds, load_result_logs
 from .load_meta_log import load_meta_log, subselect_meta_log
 from .load_hyper_log import load_hyper_log, subselect_hyper_log
 from .load_model import load_model_ckpt
-
+from .helpers import (print_framed,
+                      get_jax_os_ready,
+                      save_pkl_object)
 
 __all__ = [
            'load_mle_toolbox_config',